# Conflicts:
#	.gradle/8.3/executionHistory/executionHistory.bin
#	.gradle/8.3/executionHistory/executionHistory.lock
#	.gradle/8.3/fileHashes/fileHashes.bin
#	.gradle/8.3/fileHashes/fileHashes.lock
#	.gradle/8.3/fileHashes/resourceHashesCache.bin
#	.gradle/file-system.probe
#	.idea/workspace.xml
#	build/tmp/compileJava/previous-compilation-data.bin
#	src/main/java/Unknown/PreScent/dto/FinishedProductDto.java
#	src/main/java/Unknown/PreScent/entity/FinishedProductEntity.java

+++ src/main/java/Unknown/PreScent/dto/SellerDto.java
@@ -12,7 +12,7 @@ import lombok.ToString;
 @ToString
 public class SellerDto {
 
-    private Long sellerKey;    //system ID
+    private Integer sellerKey;    //system ID
     private String sellerName;
     private String sellerId;
     private String sellerPassword;
